pyproject.toml
[project]
name = "icp-query"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.3"
dependencies = [
    "aiosqlite>=0.21.0",
    "ddddocr>=1.5.6",
    "fastapi[standard]>=0.115.12",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "imagehash>=4.3.2",
    "numpy>=2.2.4",
    "onnxruntime>=1.21.0",
    "opencv-python-headless>=4.11.0.86",
    "pycryptodome>=3.22.0",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.12.0",
    "python-json-logger>=4.0.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.34.2",
    "verboselogs>=1.7",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.6",
    "isort>=6.0.1",
    "matplotlib>=3.10.1",
    "pre-commit>=4.2.0",
    "pyinstrument>=5.1.1",
    "rich>=14.2.0",
    "ruff>=0.11.6",
    "scalene>=1.5.55",
    "scikit-learn>=1.6.1",
    "tqdm>=4.67.1",
]
